requirements.txt
fastapi
uvicorn
websockets
faster-whisper
pydantic
pydantic-settings
python-dotenv
faiss-cpu
sentence-transformers
pymupdf4llm
markitdown
groq
httpx
python-multipart